Move Indexes and Primary index into domain Closes #490
Move Indexes and Primary index into domain

Closes #490
