+++ v0.4/platform/core/src/classes.ts
@@ -111,10 +111,6 @@ export interface ArrayOf extends Type {
 
 ///
 
-export interface Indices extends Mixin<Doc> {
-  primary: string
-}
-
 export interface AnyLayout {
   [key: string]: PropertyType
 }
@@ -139,5 +135,3 @@ export const CORE_CLASS_REF_TO = 'class:core.RefTo' as Ref<Class<RefTo<Doc>>>
 export const CORE_CLASS_BAG_OF = 'class:core.BagOf' as Ref<Class<BagOf>>
 export const CORE_CLASS_INSTANCE_OF = 'class:core.InstanceOf' as Ref<Class<InstanceOf<Emb>>>
 export const CORE_CLASS_ENUM_OF = 'class:core.EnumOf' as Ref<Class<EnumOf<EnumKey>>>
-
-export const CORE_MIXIN_INDICES = 'mixin:core.Indices' as Ref<Mixin<Indices>>
