+++ v0.4/platform/core/src/ids.ts
@@ -1,0 +1,28 @@
+///
+
+import { Doc, Ref } from './classes'
+
+function toHex (value: number, chars: number): string {
+  const result = value.toString(16)
+  if (result.length < chars) {
+    return '0'.repeat(chars - result.length) + result
+  }
+  return result
+}
+
+let counter = Math.random() * (1 << 24) | 0
+const random = toHex(Math.random() * (1 << 24) | 0, 6) + toHex(Math.random() * (1 << 16) | 0, 4)
+
+function timestamp (): string {
+  const time = (Date.now() / 1000) | 0
+  return toHex(time, 8)
+}
+
+function count (): string {
+  const val = counter++ & 0xffffff
+  return toHex(val, 6)
+}
+
+export function generateId (): Ref<Doc> {
+  return timestamp() + random + count() as Ref<Doc>
+}
