+++ v0.4/platform/core/src/index.ts
@@ -16,3 +16,4 @@
 export * from './classes'
 export * from './model'
 export * from './storage'
+export * from './ids'
